🎉 Release 1.9.0 (#182)
This PR was opened by the [ready-release-go](https://github.com/woodpecker-ci/plugin-ready-release-go) plugin. When you're ready to do a release, you can merge this pull-request and a new release with version `1.9.0` will be created automatically. If you're not ready to do a release yet, that's fine, whenever you add more changes to `master` this pull-request will be updated.

## Options

- [ ] Mark this version as a release candidate

## [1.9.0](https://gitea.ocram85.com/arkanum/arkanum/releases/tag/v1.9.0) - 2025-05-08

### ✨ Features

- bump version because of 2 skipped code-server releases [[#181](https://gitea.ocram85.com/arkanum/arkanum/pulls/181)]

Reviewed-on: https://gitea.ocram85.com/arkanum/arkanum/pulls/182
Reviewed-by: OCram85 <marco.blessing@googlemail.com>
Co-authored-by: oauth <noreply@ocram85.com>
Co-committed-by: oauth <noreply@ocram85.com>